+++ src/main/java/controllers/HelloController.java
@@ -1,4 +1,4 @@
-package com.client.client;
+package controllers;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
